Replace undefined in rootstate to null to support JSON
