+++ web/src/StateStore/_gen/longNote_state.ts
@@ -3,8 +3,8 @@ import { LongNote } from '~NoteView/types';
 
 export class LongNoteState extends Record<{
   longNotes: List<LongNote>;
-  editingLongNote?: Partial<LongNote>
+  editingLongNote: Partial<LongNote> | null
 }>({
   longNotes: List(),
-  editingLongNote: undefined,
+  editingLongNote: null,
 }) {};
